+++ synth/generator/utils.go
@@ -9,9 +9,9 @@ package generator
 type Pwm struct {
 }
 
-func RescaleMidiValues(value int64, outMin float64, outMax float64) float64 {
-	return float64(value-0)*(outMax-outMin)/float64(127) + outMin
+func RescaleMidiValues(value int64, outMin float32, outMax float32) float32 {
+	return float32(value-0)*(outMax-outMin)/float32(127) + outMin
 }
-func RescaleToMilliSeconds(value int32, inMin float64, inMax float64, milisecondsAmount int32) float64 {
-	return (float64(value) - inMin) * float64(milisecondsAmount*1000) / inMax
+func RescaleToMilliSeconds(value int32, inMin float32, inMax float32, milisecondsAmount int32) float32 {
+	return (float32(value) - inMin) * float32(milisecondsAmount*1000) / inMax
 }
